Middleware implemented

+++ controllers/NewArrival.controller.js
@@ -3,6 +3,16 @@ const NewArrival = mongoose.model('NewArrival');
 
 const enterNewArrival = async (req, res) => {
     try {
+        //This gets the object with the information that we passed, like idClient, rol and IAT
+        //The "iat" (issued at) claim identifies the time at which the JWT was issued, it uses to 
+        //implement expiration time or date for a token (for security purporse)
+        //console.log(req.user);
+        if(req.user.rol !== 'admin') {
+            return res.status(401).json({
+                msg: 'Unauthorized',
+                details: 'You do not have access to make this process',
+            });
+        }
         const { name, features, price, availability, quantity, costPrice, itemImage } = req.body;
 
         const newElement = {
@@ -70,6 +80,14 @@ const getNewArrivalById = async (req, res) => {
 
 const updateNewArrival = async (req, res) => {
     try {
+
+        if(req.user.rol !== 'admin') {
+            return res.status(401).json({
+                msg: 'Unauthorized',
+                details: 'You do not have access to make this process',
+            });
+        }
+
         const { id } = req.params;
 
         const result = await NewArrival.findByIdAndUpdate(
@@ -93,6 +111,14 @@ const updateNewArrival = async (req, res) => {
 
 const removeNewArrival = async (req, res) => {
     try {
+
+        if(req.user.rol !== 'admin') {
+            return res.status(401).json({
+                msg: 'Unauthorized',
+                details: 'You do not have access to make this process',
+            });
+        }
+
         const { id } = req.params;
 
         const result = await NewArrival.findByIdAndDelete(id);
